==> notebooks/00_chembl_go_query.ipynb <==
import pandas as pd
import sqlite3

conn = sqlite3.connect('../data/chembl_33.db')

cursor = conn.execute("SELECT name FROM sqlite_master WHERE type='table';")
try:
    assert len(cursor.fetchall()) > 1
except AssertionError:
    print('Incorrect database. Please download the database again.')

_sql = """
SELECT
    MOLECULE_DICTIONARY.pref_name,
    MOLECULE_DICTIONARY.CHEMBL_ID as chembl_idx,
    ASSAYS.chembl_id as assay_id,
    ACTIVITIES.pchembl_value as pchembl_value,
    GO_CLASSIFICATION.go_id as go_id,
    GO_CLASSIFICATION.pref_name as go_term
FROM MOLECULE_DICTIONARY
JOIN ACTIVITIES ON MOLECULE_DICTIONARY.molregno == ACTIVITIES.molregno
JOIN ASSAYS ON ACTIVITIES.assay_id == ASSAYS.assay_id
JOIN TARGET_DICTIONARY on ASSAYS.tid == TARGET_DICTIONARY.tid
JOIN TARGET_COMPONENTS on TARGET_DICTIONARY.tid == TARGET_COMPONENTS.tid
JOIN COMPONENT_SEQUENCES on TARGET_COMPONENTS.component_id == COMPONENT_SEQUENCES.component_id
JOIN COMPONENT_GO on COMPONENT_SEQUENCES.component_id == COMPONENT_GO.component_id
JOIN GO_CLASSIFICATION on COMPONENT_GO.go_id == GO_CLASSIFICATION.go_id
JOIN COMPOUND_STRUCTURES on MOLECULE_DICTIONARY.molregno == COMPOUND_STRUCTURES.molregno
WHERE
    ASSAYS.tid is not null
    and ASSAYS.assay_organism = 'Homo sapiens'
"""

df = pd.read_sql(_sql, con=conn)
df.head()

df.dropna(subset=['pchembl_value'], inplace=True)
df.shape

df.to_csv('../data/mappings/chembl_go.tsv', sep='\t', index=False)
