==> notebooks/0_preprocessing.ipynb <==
# Notebook to extract base networks from raw data.

import pandas as pd
from tqdm import tqdm

KG_DIR = '../data/kg'

# Load raw kg

Loading the KG from [drug2ways](https://github.com/drug2ways/results/tree/master) resource:

First, the [OpenBioLink KG](https://github.com/OpenBioLink/OpenBioLink):

openbiolink_df = pd.read_csv(
    '../data/kg/raw/openbiolink_filtered_kg.tsv', sep='\t'
)
openbiolink_df.head(2)

... and the Custom KG, or "In-House" KG, described in the [drug2ways paper](https://journals.plos.org/ploscompbiol/article/file?id=10.1371/journal.pcbi.1008464&type=printable).

custom_df = pd.read_csv(
    '../data/kg/raw/custom_filtered_kg.tsv', sep='\t'
)
custom_df.head(2)

Let's investigate the columns here:

custom_df['polarity'].unique()

Seems that polarity indicates directionality here.

custom_df.head(50)

While I can only see genes/proteins and drugs, the original paper says that this KG also contains phenotypes and diseases/indications.

# Subsetting only drug-protein and protein-protein edges

These KGs have PUBCHEM.COMPOUND ids as drugs and NCBIGENE IDs as proteins.

# drug-protein
openbiolink_drug_edges = openbiolink_df[
    (openbiolink_df['source'].str.contains('pubchem.compound')) & (openbiolink_df['target'].str.contains('ncbigene'))
]

# protein-protein
openbiolink_pp_edges = openbiolink_df[
    (openbiolink_df['source'].str.contains('ncbigene')) & (openbiolink_df['target'].str.contains('ncbigene'))
]

openbiolink_drug_edges.shape, openbiolink_pp_edges.shape

# drug-protein
custom_drug_edges = custom_df[
    (custom_df['source'].str.contains('pubchem.compound')) & (custom_df['target'].str.contains('ncbigene'))
]

# protein-protein
custom_pp_edges = custom_df[
    (custom_df['source'].str.contains('ncbigene')) & (custom_df['target'].str.contains('ncbigene'))
]

custom_drug_edges.shape, custom_pp_edges.shape

# Saving the subset graph

# concatenate the drug-protein and protein-protein edges for each respective KG
openbiolink_filtered = pd.concat([openbiolink_drug_edges, openbiolink_pp_edges])
custom_filtered = pd.concat([custom_drug_edges, custom_pp_edges])

# save each to separate files
openbiolink_filtered.to_csv(f'{KG_DIR}/openbiolink_filtered.tsv', sep='\t', index=False)
custom_filtered.to_csv(f'{KG_DIR}/customkg_filtered.tsv', sep='\t', index=False)
