==> notebooks/3_uniprot_keywords.ipynb <==
# Explore UniProt keywords

url: https://www.uniprot.org/uniprotkb?facets=reviewed%3Atrue&query=%28taxonomy_id%3A9606%29
1. Select the following columns:
Entry	Entry Name	Protein Names	Gene Names	Organism	Keywords HGNC

2. Select Reviewed only (~20,500)

from collections import Counter

import pandas as pd
from tqdm import tqdm
import seaborn as sns
import matplotlib.pyplot as plt
import itertools as it

import os.path as osp
import json

Below, I downloaded the [UniProt table for human proteins](https://www.uniprot.org/uniprotkb?query=%28Human%29). I added on Keywords and HGNC as columns:

MAPPING_DIR = '../data/mappings/'

# Release 2022_05
uniprot = pd.read_csv(
    osp.join(MAPPING_DIR, 'uniprotkb_Human_AND_reviewed_true_2023_10_05.tsv.gz'),
    sep='\t',
    compression='gzip',
    usecols=[
        'Keywords',
        'Entry',
    ]
)

uniprot.head(10)

Get our KG and all the protein nodes inside. We are getting the KG which includes the test set:

KG_DIR = '../data/kg'

kg = pd.read_csv(osp.join(KG_DIR, 'final_kg.tsv'), sep='\t')

Get the proteins:

target_sets = set(kg[kg['target'].str.startswith('ncbigene')]['target'].tolist())
source_sets = set(kg[kg['source'].str.startswith('ncbigene')]['source'].tolist())
proteins_in_kg = target_sets.union(source_sets)
len(proteins_in_kg)

Next, let's get a mapping from UniProt to NCBI gene ID:

with open(osp.join(MAPPING_DIR, 'uniprot2ncbi.json')) as f:
    uniprot2ncbi = json.load(f)

Get the valid protein classes:

valid_keywords = pd.read_csv(
    osp.join(MAPPING_DIR, 'molecular_function.tsv'),
    sep='\t',
)

valid_keywords

molecular_functions = valid_keywords.Name.unique()

Now, we'll make a dictionary mapping each of our KG proteins to their protein class (or keywords):

all_keywords = []

not_mapped = 0
no_keywords = 0

protein_to_keywords = {}

for prot_name, keywords in tqdm(uniprot.values):
    # Map from UniProt to NCBI
    if 'UniProt:' + str(prot_name) not in uniprot2ncbi:
        not_mapped += 1
        continue
    
    """Get keywords"""
    protein_keywords = set()
    
    for keyword in keywords.split(';'):
        
        if keyword not in molecular_functions:
            continue
        
        protein_keywords.add(keyword.strip())
        all_keywords.append(keyword.strip())
        
    if not protein_keywords:
        no_keywords += 1
        continue
        
    protein_to_keywords[uniprot2ncbi['UniProt:' + str(prot_name)]] = protein_keywords

len(protein_to_keywords), no_keywords, not_mapped, len(set(all_keywords))

Which keywords are most common? 

valid_keywords = [
    k
    for k, c in Counter(all_keywords).items()
    if c > 50
]

len(valid_keywords)

Removing superclasses and keywords with a few proteins

protein_to_keywords = {
    protein: {
        keyword
        for keyword in keywords
        if keyword in valid_keywords and keyword not in [
            'Transducer',
            'Serine/threonine-protein kinase',
            'Activator',
        ]
    }
    for protein, keywords in protein_to_keywords.items()
}

Map the proteins in our KG to Uniprot keywords:

nodes2keywords = {key: val for key, val in protein_to_keywords.items() if key in proteins_in_kg and len(val) > 0}

Those which map to classes:

len(nodes2keywords)

Those which don't:

len(proteins_in_kg) - len(nodes2keywords)

2769 / (1268+2769)

Get some further stats:

keywords_in_nodes = list()
number_of_keywords = dict()
for key, val in nodes2keywords.items():
    keywords_in_nodes.extend(list(val))
    number_of_keywords[key] = len(val)

Let's see the most popular classes:

most_popular = Counter(keywords_in_nodes)
most_popular.most_common()

How many classes per protein?

sns.displot(list(number_of_keywords.values()))
# add x axis label
plt.xlabel("Number of protein functional keywords")
# add y axis label
plt.ylabel("Number of nodes")

From those above, we want to reduce the number of keywords. Let's see how many keywords are amongst those with just 1 keyword:

# get all the protein IDs
single_keyword = {key for key, val in number_of_keywords.items() if val == 1}

# and then the corresponding keywords to those proteins
kws = set()
for key, val in nodes2keywords.items():
    if key in single_keyword:
        kws.update(val)

# now subet nodes2keywords to include only those proteins with a single keyword
nodes2keywords = {key: val.intersection(kws) for key, val in nodes2keywords.items() if len(val.intersection(kws)) > 0}

len(nodes2keywords)

How much of the proteins in the KG does that cover?

len(nodes2keywords) / len(proteins_in_kg)

len(kws)

We'll cut these down a bit more to start small:

First, get all the proteins connected to drugs or BPS:

prots2bps = set(kg.loc[kg['edge_type'] == 'participates']['source'])

prots_frm_drugs = set(kg.loc[(kg['edge_type'] == 'upregulates') | (kg['edge_type'] == 'downregulates')]['target'])

critical_prots = prots2bps | prots_frm_drugs

# and then the corresponding keywords to those proteins
critical_kws = list()
kw_freqs = list()
for key, val in nodes2keywords.items():
        for word in val:
            if key in critical_prots:
                critical_kws.append(word)
            kw_freqs.append(word)

critical_counts = Counter(critical_kws)
kw_freqs_counts = Counter(kw_freqs)

Now, we see what percentage of each class comprises vital proteins:

percentages = dict()
for key, val in kw_freqs_counts.items():
    if key in critical_counts:
        percentages[key] = critical_counts[key] / val
    else:
        percentages[key] = 0

percentages = Counter(percentages).most_common()

percentages

kw_freqs_counts.most_common()

The code below gets us a good amount, but we'll start with a smaller number to test it out:

# keep_kws should be the words in percentages corresponding to those with >80% value
#keep_kws = {key for key, val in percentages if val > 0.80 or kw_freqs_counts[key] > 150}
#len(keep_kws)

# keep_kws should be the words in percentages corresponding to those with >80% value
keep_kws = {key for key, val in percentages if val > 0.5}
#keep_kws = {key for key, val in kw_freqs_counts.items() if val > 1000}
len(keep_kws)

Okay, how many do we have left?

# now subet nodes2keywords to include only those proteins with a single keyword
nodes2keywords = {key: val.intersection(keep_kws) for key, val in nodes2keywords.items() if len(val.intersection(keep_kws)) > 0}

len(nodes2keywords) / len(proteins_in_kg)

We still have about a quarter left.

So if we narrow it down to these classes, we still include all the proteins we had before.

# get most_popular for only terms for kws
keyword_freqs = {key: val for key, val in most_popular.items() if key in keep_kws}
keyword_freqs

Okay, let's get a dictionary mapping proteins in the KG to their class. If they don't map to anything, map them to a generic protein class ('Gene', to keep it consistent with the KG).

node2class = dict()
no_class_nodes = set()

for node in proteins_in_kg: # take away non-critical proteins that don't have a keyword
    if node in nodes2keywords:
        node2class[node] = list(nodes2keywords[node])
    elif node in critical_prots:  
        #node2class[node] = ['Gene']
        no_class_nodes.add(node)

# get rid of the ones that are neither critical nor have any keywords
nodes2remove = set(proteins_in_kg) - set(node2class.keys())
len(nodes2remove)

# write dictionary to json file in MAPPING_DIR
with open(osp.join(MAPPING_DIR, 'node2class.json'), 'w') as f:
    json.dump(node2class, f, indent=4)

## Alter the KG:

Now, we want to alter the KG so that we replace rows with the interactions from those limited classes.

kg.head()

Change the protein classes:

new_rows = []

Now we add to our new KG:

for i, row in tqdm(kg.iterrows()):
    # first, check whether the nodes are within our node classes:
    if row['source'] in nodes2remove or row['target'] in nodes2remove:
        continue
    # if both match:
    elif row['source'] in node2class and row['target'] in node2class:
        for combo in list(it.product(node2class[row['source']], node2class[row['target']])):
            s_kw, t_kw = combo
            kg.loc[i, 'source_node_type'] = s_kw
            kg.loc[i, 'target_node_type'] = t_kw
            new_rows.append(list(row))
    # if just the source is in our dict
    elif row['source'] in node2class:
        for kw in node2class[row['source']]:
            kg.loc[i, 'source_node_type'] = kw
            new_rows.append(list(row))
    # if just the target is in our dict
    elif row['target'] in node2class:
        for kw in node2class[row['target']]:
            kg.loc[i, 'target_node_type'] = kw
            new_rows.append(list(row))
    else:
        new_rows.append(list(row))

kg_subclass = pd.DataFrame(new_rows, columns=kg.columns)

kg_subclass.head()

kg_subclass['source_node_type'].value_counts()

kg_subclass['edge_type'].value_counts()

kg_subclass.to_csv(osp.join(KG_DIR, 'final_kg_subclassed.tsv'), sep='\t', index=False)
