==> notebooks/4_train_test_splits.ipynb <==
# Train - Test - Validation Splits

import pandas as pd
import os.path as osp
import os
import networkx as nx
from copy import deepcopy

## Load in the final KGs

KG_DIR = '../data/kg'

kg = pd.read_csv(osp.join(KG_DIR, 'final_kg.tsv'), sep='\t')
kg.drop_duplicates(inplace=True)

kg_protclass = pd.read_csv(osp.join(KG_DIR, 'final_kg_subclassed.tsv'), sep='\t')
kg_protclass.drop_duplicates(subset=['source', 'target'], inplace=True)

drug_bp_pairs = kg.loc[kg['edge_type'] == 'induces']

drug_bp_pairs_protclass = kg_protclass.loc[kg_protclass['edge_type'] == 'induces']

print(f"There are {len(drug_bp_pairs)} positive drug-BP pairs in the final KG")

print(f"There are {len(drug_bp_pairs_protclass)} positive drug-BP pairs in the protein-classed KG")

## Load in the DrugMechDB pairs that go in the test set

dm_db_pairs = pd.read_csv(osp.join(KG_DIR, 'drugmechdb_triples.tsv'), sep='\t')
dm_db_pairs.drop_duplicates(inplace=True)

print(f"{len(dm_db_pairs)} additional drug-BP pairs come from DrugMechDB, constituting {len(dm_db_pairs)/len(drug_bp_pairs)*100}%")

## Specify which pairs are able to be matched:

MARS / MARS can match pairs as long as they are 4 or less hops apart (our hyperparameter setting) and not connected via an inverse _CtBP edge.

G = nx.DiGraph()

for i, row in kg.iterrows():
    if row['edge_type'] == 'induces':
        continue
    src_id = row['source']
    trgt_id = row['target']
    if src_id not in G.nodes:
        G.add_node(src_id, type=row['source_node_type'])
    if trgt_id not in G.nodes:
        G.add_node(trgt_id, type=row['target_node_type'])
    G.add_edge(src_id, trgt_id, type=row['edge_type'])

Gp = nx.DiGraph()

for i, row in kg_protclass.iterrows():
    if row['edge_type'] == 'induces':
        continue
    src_id = row['source']
    trgt_id = row['target']
    if src_id not in Gp.nodes:
        Gp.add_node(src_id, type=row['source_node_type'])
    if trgt_id not in Gp.nodes:
        Gp.add_node(trgt_id, type=row['target_node_type'])
    Gp.add_edge(src_id, trgt_id, type=row['edge_type'])

Check the unmatched pairs for the regular, full KG:

unmatched_pairs = set()
unmatched_dm_db_pairs = set()

path_lengths = dict()

for i, row in drug_bp_pairs.iterrows():
    if not nx.has_path(G, row['source'], row['target']):
        unmatched_pairs.add(i)
    elif 1 > nx.shortest_path_length(G, row['source'], row['target']) > 4:
        unmatched_pairs.add(i)
    else:
        path_lengths[f"{(row['source'], row['target'])}"] = nx.shortest_path_length(G, row['source'], row['target'])

for i, row in dm_db_pairs.iterrows():
    if not nx.has_path(G, row['source'], row['target']):
        unmatched_dm_db_pairs.add(i)
    elif 1 > nx.shortest_path_length(G, row['source'], row['target']) > 4:
        unmatched_dm_db_pairs.add(i)

print(f"{len(unmatched_pairs)} drug-BP pairs in the final KG are not connected via 4 hops or less")
print(f"{len(unmatched_dm_db_pairs)} drug-BP pairs from DrugMechDB are not connected via 4 hops or less")

Check the unmatched pairs for the protein-subclassed KG:

unmatched_pairs_p = set()
unmatched_dm_db_pairs_p = set()

path_lengths_p = dict()

for i, row in drug_bp_pairs_protclass.iterrows():
    if not row['source'] in Gp.nodes or not row['target'] in Gp.nodes:
        unmatched_pairs_p.add(i)
    elif not nx.has_path(Gp, row['source'], row['target']):
        unmatched_pairs_p.add(i)
    elif 1 > nx.shortest_path_length(Gp, row['source'], row['target']) > 4:
        unmatched_pairs_p.add(i)
    else:
        path_lengths_p[f"{(row['source'], row['target'])}"] = nx.shortest_path_length(Gp, row['source'], row['target'])

for i, row in dm_db_pairs.iterrows():
    if not row['source'] in Gp.nodes or not row['target'] in Gp.nodes:
        unmatched_dm_db_pairs_p.add(i)
    elif not nx.has_path(Gp, row['source'], row['target']):
        unmatched_dm_db_pairs_p.add(i)
    elif 1 > nx.shortest_path_length(Gp, row['source'], row['target']) > 4:
        unmatched_dm_db_pairs_p.add(i)

print(f"{len(unmatched_pairs_p)} drug-BP pairs in the protein-classed KG are not connected via 4 hops or less")
print(f"{len(unmatched_dm_db_pairs_p)} drug-BP pairs from DrugMechDB are not connected via 4 hops or less")

# Remove unmatched pairs from the final KG
drug_bp_pairs.drop(list(unmatched_pairs), inplace=True)
drug_bp_pairs.reset_index(drop=True, inplace=True)

# Remove unmatched pairs from the protein-classed KG
drug_bp_pairs_protclass.drop(list(unmatched_pairs_p), inplace=True)
drug_bp_pairs_protclass.reset_index(drop=True, inplace=True)

# Remove unmatched pairs from DrugMechDB
# make copies of it for multiple KG variants:
dm_db_pairs_protclass = deepcopy(dm_db_pairs)

dm_db_pairs.drop(list(unmatched_dm_db_pairs), inplace=True)
dm_db_pairs.reset_index(drop=True, inplace=True)

dm_db_pairs_protclass.drop(list(unmatched_dm_db_pairs_p), inplace=True)
dm_db_pairs_protclass.reset_index(drop=True, inplace=True)

Add their categories (path lengths) in:


## Create the Splits

In the original PoLo example, they use different proportions for the splits, but let's go with something most similar to their Hetionet example, in which they do an approximate 60/20/20% split.

Note that the DrugMechDB examples need to be in the test set.

First, we'll exclude the subset of DrugMechDB examples that are in the test set.

Then, we'll split the remaining examples into train, validation, and test sets, with the test set accumulating to 20% with the DrugMechDB examples.

total_positives = len(drug_bp_pairs) + len(dm_db_pairs)

total_positives_pc = len(drug_bp_pairs_protclass) + len(dm_db_pairs_protclass)

So we need to get the following numbers from the KG positive examples for train, validation, and test sets:

proportions = round(0.6 * total_positives), round(0.2 * total_positives), round(0.2 * total_positives) - len(dm_db_pairs)
print(proportions)

proportions_pc = round(0.6 * total_positives_pc), round(0.2 * total_positives_pc), round(0.2 * total_positives_pc) - len(dm_db_pairs_protclass)
print(proportions_pc)

# write a function which separates the dataframe into train, val and test sets of defined sizes
def train_test_split(df, train_size, val_size, test_size):
    df = df.sample(frac=1, random_state=7).reset_index(drop=True)
    train = df[:train_size]
    val = df[train_size:train_size+val_size]
    test = df[train_size+val_size:train_size+val_size+test_size]
    return train, val, test

train, val, test = train_test_split(drug_bp_pairs, proportions[0], proportions[1], proportions[2])

train_pc, val_pc, test_pc = train_test_split(drug_bp_pairs_protclass, proportions_pc[0], proportions_pc[1], proportions_pc[2])

Check it did what we want:

len(train), len(val), len(test)

len(train_pc), len(val_pc), len(test_pc)

No overlap?

train_pairs = {(row['source'], row['target']) for i, row in train.iterrows()}
test_pairs = {(row['source'], row['target']) for i, row in test.iterrows()}
val_pairs = {(row['source'], row['target']) for i, row in val.iterrows()}

train_pairs & test_pairs

train_pairs & val_pairs

test_pairs & val_pairs

Overlap in the protein-classed one?

train_pairs = {(row['source'], row['target']) for i, row in train_pc.iterrows()}
test_pairs = {(row['source'], row['target']) for i, row in test_pc.iterrows()}
val_pairs = {(row['source'], row['target']) for i, row in val_pc.iterrows()}

train_pairs & test_pairs

train_pairs & val_pairs

test_pairs & val_pairs

Good, no overlap. Add the DrugMechDB examples to the test set:

test = pd.concat([test, dm_db_pairs]).sample(frac=1, random_state=7).reset_index(drop=True)

test_pc = pd.concat([test_pc, dm_db_pairs_protclass]).sample(frac=1, random_state=7).reset_index(drop=True)

print(len(test))
print(len(test_pc))

Take the test and validation sets out the KG:

kg = kg.loc[kg['edge_type'] != 'induces']
kg_mars = pd.concat([kg, train]).sample(frac=1, random_state=7).reset_index(drop=True)

kg_protclass = kg_protclass.loc[kg_protclass['edge_type'] != 'induces']
kg_mars_pc = pd.concat([kg_protclass, train_pc]).sample(frac=1, random_state=7).reset_index(drop=True)

len(kg_mars.loc[kg_mars['edge_type'] == 'induces']) == len(train)

len(kg_mars_pc.loc[kg_mars_pc['edge_type'] == 'induces']) == len(train_pc)

Write everything to files:

SPLITS_DIR = osp.join(KG_DIR, 'splits')
MOA_NET = osp.join(SPLITS_DIR, 'MoA-net')
MOA_NET_PROTCLASSED = osp.join(SPLITS_DIR, 'MoA-net-protclass')

# create the above directories if they do not exist:
for dir in [SPLITS_DIR, MOA_NET, MOA_NET_PROTCLASSED]:
    if not osp.exists(dir):
        os.makedirs(dir)

kg.to_csv(osp.join(MOA_NET, 'kg_no_cmp_bp.tsv'), sep='\t', index=False)
kg_mars.to_csv(osp.join(MOA_NET, 'kg_with_train_smpls.tsv'), sep='\t', index=False)

train.to_csv(osp.join(MOA_NET, 'train.tsv'), sep='\t', index=False)
val.to_csv(osp.join(MOA_NET, 'dev.tsv'), sep='\t', index=False)
test.to_csv(osp.join(MOA_NET, 'test.tsv'), sep='\t', index=False)

kg_protclass.to_csv(osp.join(MOA_NET_PROTCLASSED, 'kg_no_cmp_bp.tsv'), sep='\t', index=False)
kg_mars_pc.to_csv(osp.join(MOA_NET_PROTCLASSED, 'kg_with_train_smpls.tsv'), sep='\t', index=False)

train_pc.to_csv(osp.join(MOA_NET_PROTCLASSED, 'train.tsv'), sep='\t', index=False)
val_pc.to_csv(osp.join(MOA_NET_PROTCLASSED, 'dev.tsv'), sep='\t', index=False)
test_pc.to_csv(osp.join(MOA_NET_PROTCLASSED, 'test.tsv'), sep='\t', index=False)

Finally, let's make an ablation test set for MoA-net to ensure the improved metrics aren't just coming from a smaller test set:

sub_test = test.sample(100, random_state=7)

sub_test.to_csv(osp.join(MOA_NET, 'sub_test.tsv'), sep='\t', index=False)
